BeeVote/requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
python-dotenv
drf-nested-routers